+++ client/components/index.js
@@ -11,3 +11,4 @@ export {default as Main} from './mainComponent/main'
 export {default as AddFriend} from './addFriend'
 export {default as UpdateForm} from './updateUserForm'
 export {default as Welcome} from './welcome'
+export {default as SearchFriend} from './searchFriend'
