+++ client/components/mainComponent/messages.js
@@ -1,3 +1,4 @@
+/* eslint-disable */
 import React, {Component} from 'react'
 import {
   getAllMessages,
