+++ package.json
@@ -49,6 +49,7 @@
     "@fortawesome/react-fontawesome": "^0.1.13",
     "@google-cloud/storage": "^5.5.0",
     "@google-cloud/translate": "^6.0.4",
+    "@google-cloud/vision": "^2.1.2",
     "axios": "^0.19.0",
     "bootstrap": "^4.5.3",
     "compression": "^1.7.3",
